Add readthedocs and sphinx to cladetime
This is cladetime's MVP documentation. Most of the changes are
in the docs/ folder which contains readthedocs and sphinx
configuration, as well as the documentation pages written
in restructured text.

Out of scope: hosting on GitHub pages, CI/CD publishing
and previews.  The goal of this change is to give early
users enough information to get started.