Remove NCBI-based compoments from cladetime
The assign_clades CLI will no longer get its sequence data
from the NCBI API. Instead, it will use methods from the
newer CladeTime class to retrieve sequence data and metadata
from Nextstrain.

This changeset cleans up the NCBI compoments from the library
and mocks out a new process for the assign_clades CLI. Once
the changes are merged, the assign_clades CLI won't actually
do anything, but the scaffolding will be in place to build
out a newer version based on the building blocks that we're
adding to the CladeTime and Tree classes.