Remove Docker from the project setup
Docker has been working nicely for our project setup and GitHub
actions, but this changeset removes it.

After a bit of thrashing about the best way to invoke the
Nextclade CLI from this package, we landed on using their
Docker image (https://github.com/reichlab/cladetime/pull/39)

To avoid getting into the business of invoke Docker commands
within a Docker container, it will be cleaner to avoid
recommended its use for dev purposes.

The plan in a future PR is to move the project setup instructions
from the main README to a CONTRIBUTING.md and point people
to uv for local setup (which will hopefully address some of our
original reasons for dockerizing the project, e.g., preventing
the need to install Python)

This is a temporary step, because uv is looking for uv.lock
Will fix this up when addressing #40