Add a filter_sequence_data function to sequence.py
This new function downloads sequence data and filters it against
a set of sequence IDs (in our case, GenBank access numbers). The
smaller, filtered list of sequences is written to disk.

The ultimate purpose of creating this smaller sequence list is to
use it as an input for the nextclade CLI that makes clade assignments.