Remove column validations from summarize_clades (#60)
Summarize_clades was doing a collect_schema operation to ensure
that items in the group_by paramater exist as colums in the
sequence metadata. However, testing on a lower-memory laptop
revealed that collect_schema was too memory intensive to
introduce in the middle of LazyFrame handling, so we'll
take it out.