Add ability to get metadata_url for .xz files on S3
CladeTime mostly doesn't care about lzma-compressed metadata
files on S3, because it works with the full set of metadata,
which is available in the smaller and more efficient ZTSD-compressed
files.

However, for unit testing, we use the Nextstrain's 100k datasets,
and those are only published on S3 in .xz format. This changeset
allows CladeTime to open lzma (.xz) files via http by getting
the file as bytes from the requests library and passing the
bytes object to lzma.

This will allow CladeTime to run in an optional demo mode
using the smaller 100k dataset.