Add a demo mode to Cladetime (#61)
* Add ability to get metadata_url for .xz files on S3

CladeTime mostly doesn't care about lzma-compressed metadata
files on S3, because it works with the full set of metadata,
which is available in the smaller and more efficient ZTSD-compressed
files.

However, for unit testing, we use the Nextstrain's 100k datasets,
and those are only published on S3 in .xz format. This changeset
allows CladeTime to open lzma (.xz) files via http by getting
the file as bytes from the requests library and passing the
bytes object to lzma.

This will allow CladeTime to run in an optional demo mode
using the smaller 100k dataset.

* Add a "demo_mode" to CladeTime

It's difficult to do live demos of CladeTime because it's
designed to operatate against Nextstrain's full set of sequence
and sequence metadata. This changeset adds an option that will
use a smaller dataset instead:
https://docs.nextstrain.org/projects/ncov/en/latest/reference/remote_inputs.html#k-subsamples

The "demo mode" is controlled by an environment variable
CLADETIME_DEMO