Update test fixture used to simulate AWS S3
Originally, the s3_setup fixture created in conftest.py
was designed to unit test cladetime's ability to pull
the correct versionId of S3 objects when provided with
a specific date. Thus, the content of the objects was
irrelevant.

Since then, we've added Cladetime features that also require
testing the content of the files on S3. This changeset
updates the s3_setup pytest fixture to include realistic
metadata, sequence, and ncov_metadata files. Rather than
using file content to test the version, we can now
check a metadata field called "version".