Don't use moto fixture when directly access URLs via Polar
This seems obvious in hindsignt, but for .zst files, the
sequence.get_metadata function uses polars to access URLs
directly (via scan_csv). Polars uses fsspec to open remote
files, so if we pass it a url to a mock, moto-created S3
bucket, it will simply try to access a real S3 bucket
(hence the 403 errors)

The moto setup works for .xz files, because in that case,
the actual file-handling is done by requests, which then
feeds the data to polars.