Bsweger/expand moto fixtures (#66)
* Update test fixture used to simulate AWS S3

Originally, the s3_setup fixture created in conftest.py
was designed to unit test cladetime's ability to pull
the correct versionId of S3 objects when provided with
a specific date. Thus, the content of the objects was
irrelevant.

Since then, we've added Cladetime features that also require
testing the content of the files on S3. This changeset
updates the s3_setup pytest fixture to include realistic
metadata, sequence, and ncov_metadata files. Rather than
using file content to test the version, we can now
check a metadata field called "version".

* Don't use moto fixture when directly access URLs via Polar

This seems obvious in hindsignt, but for .zst files, the
sequence.get_metadata function uses polars to access URLs
directly (via scan_csv). Polars uses fsspec to open remote
files, so if we pass it a url to a mock, moto-created S3
bucket, it will simply try to access a real S3 bucket
(hence the 403 errors)

The moto setup works for .xz files, because in that case,
the actual file-handling is done by requests, which then
feeds the data to polars.

* Add more checks to sequence metadata obtained via URL

These additional checks do some basic asserts to ensure
the schema of the metadata columns used by Cladetime
and to ensure completeness/uniqueness of the strain
column (which acts, essentially, as a primary key)