Add ability to decompress ZSTD-compressed files
Nextstrain publishes the full set of SAR-CoV-2 sequences
in both lzma and zstd compression formats. We've been
using the former, because Python has a built-in
module for lzma decompression.

However, the .zst version of the file is less than half
the size of the .xz version, so this changeset incorporates
a third-party zstd binding and allows us to work with
the ZSTD compression format.