Add a workflow for building and publishing to test PyPI
Cladetime has been successfully published to test-pypi:
https://test.pypi.org/project/cladetime/

This commit adds a workflow to automate publishing to test-pypi. Once this
is hooked up, we can focus on what deployment to the real PyPI should look like
(presumably we'd want to use a similar process to what Zhian set up for our
R packages).