Additional configuration for setuptools.scm
Out of the box, setuptools_scm generates development
package versions in a format that doesn't comply with
PEP 440, so the test-pypi publish workflow failed.

That's by design, to prevent dev releases going to PyPI.
However, our prod PyPI workflow will trigger only when
a tag in the format v[0-9].[0-9].[0-9] is pushed to
GitHub, so the liklihood of inadvertently publishing
a dev version to PyPI is low.