Add dependabot dependency scanning for GitHub actions
This commmit also switches a few third-party actions specifying versions
via commit hash instead of tag (per security recommendations).

The commit hashes tag slightly older version of the actions so we can
verify whether or not dependabot will flag them as outdated (the docs
are not clear on this matter).