Do some security-related housekeeping (#104)
* Update format of comments that specify gh action version numbers

This follows the same pattern used by the variant-nowcast-hub workflows, in
the hopes that Cladetime will have similar success with Dependabot submitting
tidy PRs when it detects a new action version.

* Update pre-commit tool versions

* Update all Python requirements

This is in preparation for turning on Dependabot for Python dependencies.
The changeset also resolves two security vulnerabilities related to Jinja2

* Configure GitHub Dependabot to check Python dependencies