Manually update Python package dependencies (#127)
* Remove the version-strategy

It wasn't super clear in the docs, but looks like pip doesn't support
the "widen" version strategy:
https://github.com/dependabot/dependabot-core/discussions/10677

Also be explicit in the Dependabot configuration that we want all packages
to be considered for update (this may be useful if/when we switch the
dependabot package-ecosystem to uv)

* Remove version contstraint on Sphinx

Sphinx is now on version 8.x. We can't use the newer version due to
a conflict with myst-parser/docutils/awscli, but when that glorious
day of upgrading is upon us, let's be ready.

* Update Python dependencies

Am still working through the correct incantation/settings for getting
Dependabot to stop submitting updates that cause dependency conflicts.
This changeset reflects updates that were generated locally using the
--upgrade option for "uv pip compile"